Moved remote synchronization to be part of the Main library instead of a seperate library
