--- Duplicati/Library/RemoteSynchronization/Duplicati.Library.RemoteSynchronization.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>disable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <RootNamespace>RemoteSynchronization</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
-    <PackageReference Include="System.CommandLine.NamingConventionBinder" Version="2.0.0-beta4.22272.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Backends\Duplicati.Library.Backends.csproj" />
-    <ProjectReference Include="..\Common\Duplicati.Library.Common.csproj" />
-    <ProjectReference Include="..\Interface\Duplicati.Library.Interface.csproj" />
-    <ProjectReference Include="..\Logging\Duplicati.Library.Logging.csproj" />
-    <ProjectReference Include="..\Main\Duplicati.Library.Main.csproj" />
-    <ProjectReference Include="..\Utility\Duplicati.Library.Utility.csproj" />
-  </ItemGroup>
-
-</Project>
